fixing showing password on response

+++ controllers/clothingItems.js
@@ -1,7 +1,9 @@
 const ClothingItem = require("../models/clothingItem");
 //  const { findUser } = require("./users");
 const {
+  FORBIDDEN_ACCESS_ERROR_STATUS_CODE,
   RESOURCE_CREATED_STATUS_CODE,
+  RESOURCE_NOT_FOUND_ERROR_STATUS_CODE,
   OK_STATUS_CODE,
   errorHandling,
 } = require("../utils/errors");
@@ -22,8 +24,6 @@ const getItems = (req, res) => {
 
 const createItem = async (req, res) => {
   const { name, weather, imageUrl } = req.body;
-  /*  const userExists = await findUser(owner);
-  if (userExists) { */
   ClothingItem.create({
     name,
     weather,
@@ -36,25 +36,41 @@ const createItem = async (req, res) => {
     .catch((err) => {
       errorHandling(res, err, "Item Id");
     });
-  /*  } else {
-    return res
-      .status(RESOURCE_NOT_FOUND_ERROR_STATUS_CODE)
-      .send({ message: RESOURCE_NOT_FOUND_MESSAGE });
-  } */
 };
 
 // DELETE /items/:itemId
 
 const deleteItem = (req, res) => {
   const { itemId } = req.params;
+  const userId = req.user._id;
+  ClothingItem.findById(itemId)
+    .then((item) => {
+      if (!item) {
+        res
+          .status(RESOURCE_NOT_FOUND_ERROR_STATUS_CODE)
+          .send({ message: "The item doesnt exist." });
+        return;
+      }
+      if (item.owner.toString() !== userId) {
+        console.log(item.owner);
+        console.log(userId);
+        res
+          .status(FORBIDDEN_ACCESS_ERROR_STATUS_CODE)
+          .send({ message: "You cannot delete this item" });
+        return;
+      }
 
-  ClothingItem.findByIdAndRemove(itemId)
-    .orFail()
-    .then((clothingItem) => {
-      res.status(OK_STATUS_CODE).send({ data: clothingItem });
+      ClothingItem.findByIdAndRemove(itemId)
+        .orFail()
+        .then((clothingItem) => {
+          res.status(OK_STATUS_CODE).send({ data: clothingItem });
+        })
+        .catch((err) => {
+          errorHandling(res, err, "Item Id");
+        });
     })
     .catch((err) => {
-      errorHandling(res, err, "Item Id");
+      errorHandling(res, err, "");
     });
 };
